Added better autowiring support
