+++ test/FeaturesBundleTest.php
@@ -3,6 +3,7 @@ namespace Yannickl88\FeaturesBundle;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Yannickl88\FeaturesBundle\DependencyInjection\Compiler\AddFeaturesCompilerPass;
 use Yannickl88\FeaturesBundle\DependencyInjection\Compiler\FeaturesCompilerPass;
 
 /**
@@ -27,8 +28,12 @@ class FeaturesBundleTest extends TestCase
         $this->features_bundle->build($container);
 
         self::assertContains(
-            FeaturesCompilerPass::class,
+            AddFeaturesCompilerPass::class,
             array_map("get_class", $container->getCompilerPassConfig()->getBeforeOptimizationPasses())
+        );
+        self::assertContains(
+            FeaturesCompilerPass::class,
+            array_map("get_class", $container->getCompilerPassConfig()->getBeforeRemovingPasses())
         );
     }
 }
