+++ test/Functional/Fixtures/Dummy.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Yannickl88\FeaturesBundle\Functional\Fixtures;
+
+
+class Dummy
+{
+
+}
